backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Portfolio AI Agent Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fal-client>=0.13.1",
    "fastapi>=0.115.0",
    "fastmcp>=3.1.1",
    "httpx>=0.27.0",
    "huggingface-hub>=1.5.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.0",
    "pillow>=10.0.0",
    "pinecone>=8.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "uvicorn>=0.30.0",
]
